Add server's IP address to configuration file
wireguard_client.sh expects there to be an IP address on the last line
of the configuration file, this breaks on the initial client to be
added as there is none on the last line of the file.

This commit adds the server's expected IP address so the rest are
automatically incremented.